ChatRoom/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-socketio==5.11.0
aiosqlite==0.19.0
pydantic==2.5.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
